requirements.txt
Flask>=2.0
Flask-SQLAlchemy>=2.5
Flask-Migrate>=3.0
python-dotenv>=0.21
Flask-CORS>=3.0
PyJWT>=2.0
werkzeug>=2.0
twilio>=8.0
psycopg2-binary>=2.9
Flask-Marshmallow>=0.14